[Style] coherences .cabal and create module for integration test
