+++ GLaDOS2023.cabal
@@ -22,6 +22,8 @@ extra-source-files:
     README.md
 
 library
+    default-language: Haskell2010
+    hs-source-dirs:   lib
     exposed-modules:  Parsing,
         Parsing.Args,
         Parsing.Token,
@@ -31,13 +33,13 @@ library
         Exec.RuntimeException,
         Exec.Lookup,
         Exec.Builtins
-    hs-source-dirs:   lib
-    build-depends:    base >= 4.14.1.0, containers
-    default-language: Haskell2010
+    build-depends:    base >= 4.14.1.0,
+        containers >=0.5.10.2
 
 executable glados
+    default-language: Haskell2010
+    hs-source-dirs:   app test
     main-is:          Main.hs
-
     -- Modules included in this executable, other than Main.
     other-modules: Parsing.CptTests,
         Parsing.AstTests,
@@ -46,33 +48,42 @@ executable glados
         ExecTests
     -- LANGUAGE extensions used by modules in this package.
     -- other-extensions:
-    build-depends:    base  >= 4.14.1.0, tasty >= 1.4.3, tasty-discover >= 5.0.0, tasty-hunit >= 0.10.0.3, containers, GLaDOS2023
-    hs-source-dirs:   app test
-    default-language: Haskell2010
+    build-depends:    base  >= 4.14.1.0,
+        tasty >= 1.4.3,
+        tasty-discover >= 5.0.0,
+        tasty-hunit >= 0.10.0.3,
+        containers >=0.5.10.2,
+        GLaDOS2023
 
 test-suite unitTest
     default-language: Haskell2010
     type: exitcode-stdio-1.0
     hs-source-dirs: unitTest
+    main-is: TestMain.hs
     other-modules: Parsing.ArgsTests,
         Parsing.TokenTests,
         Parsing.CptTests,
         Parsing.AstTests,
         ParsingTests,
         ExecTests
-
-    main-is: TestMain.hs
-    build-depends: base >=4.14.1.0, tasty >= 1.4.3, tasty-discover >= 5.0.0, tasty-hunit >= 0.10.0.3, GLaDOS2023, containers
+    build-depends: base >=4.14.1.0,
+        tasty >= 1.4.3,
+        tasty-discover >= 5.0.0,
+        tasty-hunit >= 0.10.0.3,
+        containers >=0.5.10.2,
+        GLaDOS2023
 
 test-suite integrationTest
     default-language: Haskell2010
     type:               exitcode-stdio-1.0
     hs-source-dirs:     integrationTests/
     main-is:            Main.hs
+    other-modules: Print
+        Test
     build-depends:  base >=4.14.1.0,
         ansi-terminal >= 0.11.3,
-        filepath >= 1.4.2.1,
         process >= 1.6.8.0,
+        filepath >= 1.4.2.1,
         directory >= 1.3.3.0,
         exitcode >= 0.1.0.0,
         GLaDOS2023
